Moved tests/ out of mentions/ app to parent directory.
