+++ runtests.py
@@ -8,7 +8,7 @@ from django.test.utils import get_runner
 
 log = logging.getLogger(__name__)
 
-SETTINGS = "mentions.tests.config.test_settings"
+SETTINGS = "tests.config.test_settings"
 
 
 if __name__ == "__main__":
@@ -17,6 +17,6 @@ if __name__ == "__main__":
     django.setup()
 
     test_runner = get_runner(settings)()
-    failures = test_runner.run_tests(["mentions.tests"])
+    failures = test_runner.run_tests(["tests"])
 
     sys.exit(bool(failures))
