+++ tests/util/viewname.py
